+++ packages/eyes-sdk-core/test/unit/fluent/CheckSettings.spec.js
@@ -83,6 +83,20 @@ describe('CheckSettings', () => {
     })
   })
 
+  it('layoutBreakpoints', async () => {
+    const checkSettings = new CheckSettings()
+    checkSettings.layoutBreakpoints()
+    assert.deepStrictEqual(checkSettings.getLayoutBreakpoints(), true)
+    checkSettings.layoutBreakpoints(false)
+    assert.deepStrictEqual(checkSettings.getLayoutBreakpoints(), false)
+    checkSettings.layoutBreakpoints([25, 50, 100, 200])
+    assert.deepStrictEqual(checkSettings.getLayoutBreakpoints(), [25, 50, 100, 200])
+    checkSettings.layoutBreakpoints([100, 200, 200, 100, 50, 25])
+    assert.deepStrictEqual(checkSettings.getLayoutBreakpoints(), [25, 50, 100, 200])
+    checkSettings.layoutBreakpoints([])
+    assert.deepStrictEqual(checkSettings.getLayoutBreakpoints(), false)
+  })
+
   // TODO this test makes more sense to run inside docker
   it('resilient to duplicate copies of the SDK', async () => {
     const context = {